accepts float % so that iframe can be 16:9

+++ Sources/HTMLDSL/View/Dimension.swift
@@ -62,7 +62,7 @@ public enum Length: CustomStringConvertible {
 }
 
 public enum AutoInheritDimension: CustomStringConvertible {
-    case percentage(Int)
+    case percentage(Float)
     case pixel(Int)
     case auto
     case inherit
